Increase limits for code and data segments
Maximum value for the limits of code and data segments in x86 is
0xfffff, not 0xffff (it is perhaps hard to notice at first since this
value is discontinuous in the segment descriptor, with the most
significant nibble placed separately from the two least significant
bytes).

While the implementation is easy to understand, it is not feasible to
execute any more or less complex demo in current setup. Current bound is
2279 instructions and 3 registers, since with these parameters, the pointer
p1 in generate_pagetable() is equal to 0x10000000, i.e. 256 MB. This causes
the limit of data segment mentioned above to be violated (with
granularity flag set in the segment descriptor the limit is in 4KB units
and 0xffff*4KB amounts to 256MB).